Reorganising the files, now there are "two" examples, or big parts.
