+++ tutorials/00-pymapdl/readme.txt
@@ -1,3 +1,15 @@
-Tutorials Using PyMAPDL
-===========================
-These tutorials demonstrate full examples using the PyMAPDL module.
+SpaceCon 2022 PyAnsys Tutorials
+===============================
+
+These tutorials demonstrate full examples using the PyAnsys libraries.
+
+
+
+Content
+-------
+
+# To be added
+
+* (Very) short introduction to Python
+* (Very) short introduction to PyAnsys and PyMAPDL
+* (Veeery) Short example on PyMAPDL.
